+++ pkg/label/label.go
@@ -1,6 +1,8 @@
 package label
 
 const (
+	Cluster = "giantswarm.io/cluster"
+
 	Organization = "giantswarm.io/organization"
 
 	ManagedBy = "giantswarm.io/managed-by"
@@ -8,3 +10,7 @@ const (
 	// Labels, used in legacy cluster namespaces
 	LegacyCustomer = "customer"
 )
+
+type LabelsGetter interface {
+	GetLabels() map[string]string
+}